requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-filter==2.1.0
django-geojson==2.12.0
django-nose==1.4.6
django-phone-field==1.7.1
djangorestframework==3.9.4
djangorestframework-gis==0.14
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
Markdown==3.1
MarkupSafe==1.1.1
nose==1.3.7
psycopg2-binary==2.8.2
Pygments==2.4.0
python-decouple==3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
